+++ packages/timeline-state-resolver/src/integrations/vmix/__tests__/vmix.spec.ts
@@ -18,6 +18,7 @@ import {
 	MappingVMixInput,
 	MappingVMixFadeToBlack,
 	MappingVMixFader,
+	MappingVMixScript,
 } from 'timeline-state-resolver-types'
 import { ThreadedClass } from 'threadedclass'
 import { VMixDevice } from '..'
@@ -2615,6 +2616,230 @@ describe('vMix', () => {
 		expect(onFunction).toHaveBeenCalledTimes(1)
 
 		expect(onFunction).toHaveBeenNthCalledWith(1, 'SetFader', expect.stringContaining('Value=0'))
+
+		clearMocks()
+		commandReceiver0.mockClear()
+
+		await myConductor.destroy()
+
+		expect(errorHandler).toHaveBeenCalledTimes(0)
+		expect(deviceErrorHandler).toHaveBeenCalledTimes(0)
+	})
+
+	test('Script Start', async () => {
+		let device: any = undefined
+		const commandReceiver0 = jest.fn((...args) => {
+			// pipe through the command
+			return device._defaultCommandReceiver(...args)
+		})
+
+		const myLayerMapping0: MappingVMixScript = {
+			device: DeviceType.VMIX,
+			mappingType: MappingVMixType.Script,
+			deviceId: 'myvmix',
+		}
+		const myLayerMapping: Mappings = {
+			vmix_ss0: myLayerMapping0,
+		}
+
+		const myConductor = new Conductor({
+			multiThreadedResolver: false,
+			getCurrentTime: mockTime.getCurrentTime,
+		})
+		const errorHandler = jest.fn((...args) => console.log('Error in device', ...args))
+		myConductor.on('error', errorHandler)
+
+		await myConductor.init()
+
+		await runPromise(
+			myConductor.addDevice('myvmix', {
+				type: DeviceType.VMIX,
+				options: {
+					host: '127.0.0.1',
+					port: 9999,
+				},
+				commandReceiver: commandReceiver0,
+			}),
+			mockTime
+		)
+
+		const deviceContainer = myConductor.getDevice('myvmix')
+		device = deviceContainer!.device as ThreadedClass<VMixDevice>
+		const deviceErrorHandler = jest.fn((...args) => console.log('Error in device', ...args))
+		device.on('error', deviceErrorHandler)
+		device.on('commandError', deviceErrorHandler)
+
+		myConductor.setTimelineAndMappings([], myLayerMapping)
+
+		expect(mockTime.getCurrentTime()).toEqual(10050)
+		await mockTime.advanceTimeToTicks(10100)
+
+		// get initial info
+		expect(onXML).toHaveBeenCalledTimes(1)
+
+		clearMocks()
+		commandReceiver0.mockClear()
+
+		// Check that no commands has been scheduled:
+		expect(await device.queue).toHaveLength(0)
+
+		myConductor.setTimelineAndMappings([
+			{
+				id: 'ss0',
+				enable: {
+					start: 11000,
+					duration: 5000,
+				},
+				layer: 'vmix_ss0',
+				content: {
+					deviceType: DeviceType.VMIX,
+					type: TimelineContentTypeVMix.SCRIPT,
+					name: 'myscript',
+				},
+			},
+		])
+
+		await mockTime.advanceTimeToTicks(11300)
+
+		expect(commandReceiver0).toHaveBeenCalledTimes(1)
+		expect(commandReceiver0).toHaveBeenNthCalledWith(
+			1,
+			11000,
+			expect.objectContaining({
+				command: {
+					command: VMixCommand.SCRIPT_START,
+					value: 'myscript',
+				},
+			}),
+			null,
+			expect.any(String)
+		)
+
+		expect(onFunction).toHaveBeenCalledTimes(1)
+
+		expect(onFunction).toHaveBeenNthCalledWith(1, 'ScriptStart', expect.stringContaining('Value=myscript'))
+
+		clearMocks()
+		commandReceiver0.mockClear()
+
+		await myConductor.destroy()
+
+		expect(errorHandler).toHaveBeenCalledTimes(0)
+		expect(deviceErrorHandler).toHaveBeenCalledTimes(0)
+	})
+
+	test('Script Stop', async () => {
+		let device: any = undefined
+		const commandReceiver0 = jest.fn((...args) => {
+			// pipe through the command
+			return device._defaultCommandReceiver(...args)
+		})
+
+		const myLayerMapping0: MappingVMixScript = {
+			device: DeviceType.VMIX,
+			mappingType: MappingVMixType.Script,
+			deviceId: 'myvmix',
+		}
+		const myLayerMapping: Mappings = {
+			vmix_ss0: myLayerMapping0,
+		}
+
+		const myConductor = new Conductor({
+			multiThreadedResolver: false,
+			getCurrentTime: mockTime.getCurrentTime,
+		})
+		const errorHandler = jest.fn((...args) => console.log('Error in device', ...args))
+		myConductor.on('error', errorHandler)
+
+		await myConductor.init()
+
+		await runPromise(
+			myConductor.addDevice('myvmix', {
+				type: DeviceType.VMIX,
+				options: {
+					host: '127.0.0.1',
+					port: 9999,
+				},
+				commandReceiver: commandReceiver0,
+			}),
+			mockTime
+		)
+
+		const deviceContainer = myConductor.getDevice('myvmix')
+		device = deviceContainer!.device as ThreadedClass<VMixDevice>
+		const deviceErrorHandler = jest.fn((...args) => console.log('Error in device', ...args))
+		device.on('error', deviceErrorHandler)
+		device.on('commandError', deviceErrorHandler)
+
+		myConductor.setTimelineAndMappings([], myLayerMapping)
+
+		expect(mockTime.getCurrentTime()).toEqual(10050)
+		await mockTime.advanceTimeToTicks(10100)
+
+		// get initial info
+		expect(onXML).toHaveBeenCalledTimes(1)
+
+		clearMocks()
+		commandReceiver0.mockClear()
+
+		// Check that no commands has been scheduled:
+		expect(await device.queue).toHaveLength(0)
+
+		myConductor.setTimelineAndMappings([
+			{
+				id: 'ss0',
+				enable: {
+					start: 11000,
+					duration: 5000,
+				},
+				layer: 'vmix_ss0',
+				content: {
+					deviceType: DeviceType.VMIX,
+					type: TimelineContentTypeVMix.SCRIPT,
+					name: 'myscript',
+				},
+			},
+		])
+
+		await mockTime.advanceTimeToTicks(11300)
+
+		expect(commandReceiver0).toHaveBeenCalledTimes(1)
+		expect(commandReceiver0).toHaveBeenNthCalledWith(
+			1,
+			11000,
+			expect.objectContaining({
+				command: {
+					command: VMixCommand.SCRIPT_START,
+					value: 'myscript',
+				},
+			}),
+			null,
+			expect.any(String)
+		)
+
+		expect(onFunction).toHaveBeenCalledTimes(1)
+
+		expect(onFunction).toHaveBeenNthCalledWith(1, 'ScriptStart', expect.stringContaining('Value=myscript'))
+
+		await mockTime.advanceTimeToTicks(20000)
+
+		expect(commandReceiver0).toHaveBeenCalledTimes(2)
+		expect(commandReceiver0).toHaveBeenNthCalledWith(
+			2,
+			16000,
+			expect.objectContaining({
+				command: {
+					command: VMixCommand.SCRIPT_STOP,
+					value: 'myscript',
+				},
+			}),
+			null,
+			expect.any(String)
+		)
+
+		expect(onFunction).toHaveBeenCalledTimes(2)
+
+		expect(onFunction).toHaveBeenNthCalledWith(2, 'ScriptStop', expect.stringContaining('Value=myscript'))
 
 		clearMocks()
 		commandReceiver0.mockClear()